touchscreen: lge_qos: sw42902: wake the panel when longpress is enabled
Currently this device is blanking the display after doze, because
without doing so, it causes the screen to never show the AOD UI.

On caymanslm in particular, this causes weird issues with the
under-display fingerprint sensor; the touch panel is set to deep sleep
mode, so the panel never reports to userspace that the fingerprint area
has been pressed in order to unlock the phone.

This is supposed to be handled in `sw42902_lpwg_mode()` which is run on
`sw42902_suspend()`, but unfortunately userspace is trying to enable
the `longpress` sysfs far too late and the panel ends up going to deep
sleep; the check in `store_longpress()` when triggering the sysfs thus
fails.

With the preface out of the way, we can just run `sw42902_lpwg()`,
which in turn runs `sw42902_lpwg_mode()`, the latter of which will
check that longpress is enabled and re-wake the screen.

Change-Id: I7f4c82a3d5d3011986a29e3cc76a1697441f5da6
Signed-off-by: Han Sol Jin <hansol@hansol.ca>